# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  database.DimSnapshot:
    properties:
      avg_ms:
        type: number
      count:
        type: integer
      max_ms:
        type: number
      min_ms:
        type: number
      p50_ms:
        type: number
      p95_ms:
        type: number
      p99_ms:
        type: number
    type: object
  database.GlobalSnapshot:
    properties:
      avg_ms:
        type: number
      count:
        type: integer
      error_count:
        type: integer
      max_ms:
        type: number
      min_ms:
        type: number
      p50_ms:
        type: number
      p95_ms:
        type: number
      p99_ms:
        type: number
      slow_count:
        type: integer
      slow_threshold_ms:
        type: integer
    type: object
  database.HistoryPoint:
    properties:
      avg_ms:
        type: number
      count:
        type: integer
      error_count:
        type: integer
      max_ms:
        type: number
      p50_ms:
        type: number
      p95_ms:
        type: number
      p99_ms:
        type: number
      qps:
        type: number
      slow_count:
        type: integer
      timestamp:
        type: string
    type: object
  database.HistorySnapshot:
    properties:
      buckets:
        items:
          $ref: '#/definitions/database.HistoryPoint'
        type: array
      recent_qps:
        type: number
      slow_threshold_ms:
        type: integer
      step_seconds:
        type: integer
      window_seconds:
        type: integer
    type: object
  database.QueryEntry:
    properties:
      duration_ms:
        type: number
      is_error:
        type: boolean
      is_slow:
        type: boolean
      operation:
        type: string
      sql:
        type: string
      table:
        type: string
      timestamp:
        type: string
    type: object
  database.StatsSnapshot:
    properties:
      by_operation:
        additionalProperties:
          $ref: '#/definitions/database.DimSnapshot'
        type: object
      by_table:
        additionalProperties:
          $ref: '#/definitions/database.DimSnapshot'
        type: object
      global:
        $ref: '#/definitions/database.GlobalSnapshot'
      slow_queries:
        items:
          $ref: '#/definitions/database.QueryEntry'
        type: array
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: 企业级 RBAC 权限管理系统后端 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RBAC Admin API
  version: 1.0.0
paths:
...
  /monitor/sql/history:
    get:
      consumes:
      - application/json
      description: 返回指定时间窗口内按 step 聚合的时序指标(QPS/平均耗时/P50/P95/P99/最大耗时/错误/慢查询)；数据来自
        Redis 滚动窗口(3s 采样、保留 24h、跨重启),Redis 不可用时回退进程内统计
      parameters:
      - description: 时间窗口，如 5m,1h（默认 15m，范围 1m~24h）
        in: query
        name: window
        type: string
      - description: 桶粒度，如 3s,10s（默认 window/180，最小 1s）
        in: query
        name: step
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/database.HistorySnapshot'
              type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: SQL监控时序数据
      tags:
      - SQL监控
  /monitor/sql/stats:
    get:
      consumes:
      - application/json
      description: 返回 SQL 执行统计；可选 window 参数按时间窗口(如 5m,1h)聚合，留空返回累计统计
      parameters:
      - description: 时间窗口，如 5m,1h；留空返回累计统计
        in: query
        name: window
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/database.StatsSnapshot'
              type: object
        "400":
          description: window 格式无效
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: SQL监控统计
      tags:
      - SQL监控
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"